Move truncation next to other thread tests for tidy
